move Code into common.py to use `wrap_into_codes`
